Refactored shot rules
